moved monkeyType api calls out of api/pages since it's not available server side
