+++ src/services/monkey-type/get-profile.ts
@@ -1,6 +1,20 @@
-import { getMonkeyTypeProfile } from '@app/pages/api/monkey-type/get-profile';
+import { REGISTERED_USERS } from '@app/content';
 
-export const getProfile = async (username: string) => {
-  const data = await getMonkeyTypeProfile(username);
-  return data;
+export const getMonkeyTypeProfile = async (username: string) => {
+  // TODO: Type return type
+  const apiKey = REGISTERED_USERS.find(
+    (user) => user.username?.toLowerCase() === username.toLowerCase()
+  )?.apiKey;
+
+  const response = await fetch(
+    `https://api.monkeytype.com/users/${username}/profile`,
+    {
+      method: 'GET',
+      headers: {
+        'Content-Type': 'application/json',
+        Authorization: `ApeKey ${apiKey}`,
+      },
+    }
+  );
+  return response.json();
 };
